server/requirements.txt
Django==5.0.6
django-crontab==0.7.1
django-filter==24.2
django-otp==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg2==2.9.9
django-cors-headers==4.4.0
apscheduler==3.10.4
